ci(signer): add deno check gate; fix 2 pre-existing edge-function type errors
The browser-side `pnpm check` (tsc) never type-checks the edge functions —
they run on Deno with a separate module graph. The signing-service-tests
workflow ran `deno test` but not `deno check`, so type errors that don't
fire in the test path went undetected (two such errors had accumulated on
main, and this session nearly added more).

Adds a `deno check` step to signing-service-tests.yml, before Supabase
starts so it fails fast. Checking the index.ts entrypoint walks the whole
import graph (handlers -> lib -> _shared), so one check covers the function.
Also expands the workflow path triggers to include _shared/ since the
signer now imports getUserFarcasterProvider from there.

Fixes the two pre-existing errors so the new gate goes green:

1. ErrorCode union drift (accounts.ts:51 ACCOUNT_PENDING). The hand-
   maintained `ErrorCode` union in types.ts had drifted from the
   `ErrorCodes` const in errors.ts — the const had ACCOUNT_PENDING and
   several auth/validation codes the union lacked. Move the const into
   types.ts as the single source of truth, derive
   `ErrorCode = (typeof ErrorCodes)[keyof typeof ErrorCodes]`, and
   re-export ErrorCodes from errors.ts for backward compat. They can't
   drift again.

2. castAddBody.type optional (sign.ts:335). The inline type declared
   `type?: CastType` but the value is always assigned (resolvedCastType
   defaults to CAST) and @farcaster/core's makeCastAdd requires
   `type: CastType`. Make the property non-optional.

Verified: `deno check --config .../deno.json .../index.ts` exits 0
(was 2 errors). Browser tsc unchanged (17 pre-existing third-party only).
The cron (publish-cast-from-db) has its own pre-existing Deno debt and a
separate path trigger; gating it is left as a follow-up.
