+++ .github/workflows/signing-service-tests.yml
@@ -4,10 +4,12 @@ on:
   push:
     paths:
       - 'supabase/functions/farcaster-signer/**'
+      - 'supabase/functions/_shared/**'
       - '.github/workflows/signing-service-tests.yml'
   pull_request:
     paths:
       - 'supabase/functions/farcaster-signer/**'
+      - 'supabase/functions/_shared/**'
       - '.github/workflows/signing-service-tests.yml'
 
 jobs:
@@ -24,6 +26,15 @@ jobs:
       - uses: denoland/setup-deno@v1
         with:
           deno-version: v1.x
+
+      # Type-check before spinning up Supabase so type regressions fail fast
+      # and cheap. Checking the index.ts entrypoint walks the whole import
+      # graph (handlers -> lib -> _shared), so a single check covers the
+      # function. The browser-side `pnpm check` (tsc) does NOT see this code —
+      # the edge function runs on Deno with a different module graph — so this
+      # is the only gate that catches edge-function type errors.
+      - name: Type-check edge function (deno check)
+        run: deno check --config supabase/functions/farcaster-signer/deno.json supabase/functions/farcaster-signer/index.ts
 
       - name: Start Supabase
         # Exclude services the signer tests don't need. Avoids the inbucket
